+++ dev/com.ibm.ws.wsat_fat.ssl.misroute/publish/servers/WSATSSL_Client/server.xml
@@ -37,6 +37,7 @@
     recoveryGroup="recovery-group-ssl"
     recoveryIdentity="WSATSSL_Client"
     totalTranLifetimeTimeout="5m"
+    leaseExpiryThreshold="5"
   />
 
   <javaPermission codebase="${shared.resource.dir}derby/derby.jar" className="java.security.AllPermission"/>
