+++ dev/com.ibm.ws.wsat_fat.ssl.misroute/publish/servers/WSATSSL_Server1/server.xml
@@ -35,8 +35,9 @@
 
   <transaction
     backendURL="https://localhost:9444"
-	  recoveryGroup="recovery-group-ssl"
-	  recoveryIdentity="WSATSSL_Server1"
+	recoveryGroup="recovery-group-ssl"
+	recoveryIdentity="WSATSSL_Server1"
+    leaseExpiryThreshold="5"
   />
 
   <javaPermission codebase="${shared.resource.dir}derby/derby.jar" className="java.security.AllPermission"/>
