+++ dev/com.ibm.ws.wsat_fat.ssl.misroute/publish/servers/WSATSSL_Server3/server.xml
@@ -35,9 +35,10 @@
 
   <transaction
     backendURL="https://localhost:9446"
-	  recoveryGroup="recovery-group-ssl"
-	  recoveryIdentity="WSATSSL_Server3"
-	  recoverOnStartup="true"
+	recoveryGroup="recovery-group-ssl"
+	recoveryIdentity="WSATSSL_Server3"
+	recoverOnStartup="true"
+    leaseExpiryThreshold="5"
 	/>
 
   <javaPermission codebase="${shared.resource.dir}derby/derby.jar" className="java.security.AllPermission"/>
